+++ examples/bullmq-task-queue/remix.config.js
@@ -1,7 +1,0 @@
-/** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  appDirectory: "app",
-  browserBuildDirectory: "public/build",
-  publicPath: "/build/",
-  serverBuildDirectory: "build",
-};