+++ scripts/release/github.ts
@@ -34,11 +34,13 @@ export async function prsMergedSinceLastTag({
   repo,
   githubRef,
 }: PrsMergedSinceLastTagOptions): Promise<PrsMergedSinceLastTagResult> {
-  let tags = await getAllTags(owner, repo);
+  let tags = await getTags(owner, repo);
   let { currentTag, previousTag } = getPreviousTagFromCurrentTag(
     githubRef,
     tags
   );
+
+  console.log(`Getting PRs merged ${previousTag.tag}...${currentTag.tag}`);
 
   /**
     nightly > nightly => 'dev'
@@ -117,71 +119,47 @@ async function getPullRequestWithFiles(
 
 function getPreviousTagFromCurrentTag(
   currentTag: string,
-  tags: Awaited<ReturnType<typeof getAllTags>>
+  tags: Awaited<ReturnType<typeof getTags>>
 ): {
   previousTag: MinimalTag;
   currentTag: MinimalTag;
 } {
   let validTags = tags
-    .filter((tag) => {
-      // if we have a `PACKAGE_VERSION_TO_FOLLOW`
-      // we only want to get the tags related to it
-      if (PACKAGE_VERSION_TO_FOLLOW) {
-        return tag.name.startsWith(PACKAGE_VERSION_TO_FOLLOW);
-      }
-      return true;
-    })
     .map((tag) => {
       let tagName = cleanupTagName(tag.name);
       let isPrerelease = semver.prerelease(tagName) !== null;
 
-      if (!tag.commit.committer?.date) return null;
-
       return {
         tag: tagName,
-        date: new Date(tag.commit.committer.date),
+        date: new Date(tag.target.tagger.date),
         isPrerelease,
       };
     })
     .filter((v: any): v is MinimalTag => typeof v !== "undefined")
     .sort(sortByDate);
 
-  let tmpCurrentTagIndex = validTags.findIndex((tag) => tag.tag === currentTag);
-  let tmpCurrentTagInfo = validTags.at(tmpCurrentTagIndex);
-
-  if (!tmpCurrentTagInfo) {
-    throw new Error(`Could not find last tag ${currentTag}`);
-  }
-
-  let currentTagInfo: MinimalTag | undefined;
+  let currentTagIndex = validTags.findIndex((tag) => tag.tag === currentTag);
+  let currentTagInfo: MinimalTag | undefined = validTags.at(currentTagIndex);
   let previousTagInfo: MinimalTag | undefined;
 
-  // if the currentTag was a stable tag, then we want to find the previous stable tag
-  if (!tmpCurrentTagInfo.isPrerelease) {
-    let stableTags = validTags
-      .filter((tag) => !tag.isPrerelease)
-      .sort((a, b) => semver.rcompare(a.tag, b.tag));
-
-    let stableTagIndex = stableTags.findIndex((tag) => tag.tag === currentTag);
-    currentTagInfo = stableTags.at(stableTagIndex);
-    if (!currentTagInfo) {
-      throw new Error(`Could not find last stable tag ${currentTag}`);
-    }
-
-    previousTagInfo = stableTags.at(stableTagIndex + 1);
-    if (!previousTagInfo) {
-      throw new Error(`Could not find previous stable tag from ${currentTag}`);
-    }
-
-    return { currentTag: currentTagInfo, previousTag: previousTagInfo };
-  }
-
-  currentTagInfo = tmpCurrentTagInfo;
   if (!currentTagInfo) {
     throw new Error(`Could not find last tag ${currentTag}`);
   }
 
-  previousTagInfo = validTags.at(tmpCurrentTagIndex + 1);
+  // if the currentTag was a stable tag, then we want to find the previous stable tag
+  if (!currentTagInfo.isPrerelease) {
+    validTags = validTags
+      .filter((tag) => !tag.isPrerelease)
+      .sort((a, b) => semver.rcompare(a.tag, b.tag));
+
+    currentTagIndex = validTags.findIndex((tag) => tag.tag === currentTag);
+    currentTagInfo = validTags.at(currentTagIndex);
+    if (!currentTagInfo) {
+      throw new Error(`Could not find last stable tag ${currentTag}`);
+    }
+  }
+
+  previousTagInfo = validTags.at(currentTagIndex + 1);
   if (!previousTagInfo) {
     throw new Error(
       `Could not find previous prerelease tag from ${currentTag}`
@@ -235,27 +213,57 @@ async function getMergedPRsBetweenTags(
   return [...nodes, ...merged];
 }
 
-// TODO: we might be able to get away with just getting up until the "latest" tag
-async function getAllTags(owner: string, repo: string) {
-  let tags = await octokit.paginate(octokit.rest.repos.listTags, {
-    owner,
-    repo,
-  });
+interface GitHubGraphqlTag {
+  name: string;
+  target: {
+    oid: string;
+    tagger: {
+      date: string;
+    };
+  };
+}
+interface GitHubGraphqlTagResponse {
+  repository: {
+    refs: {
+      nodes: Array<GitHubGraphqlTag>;
+    };
+  };
+}
 
-  return await Promise.all(
-    tags.map(async (tag) => {
-      let commit = await octokit.rest.repos.getCommit({
-        owner,
-        repo,
-        ref: tag.commit.sha,
-      });
-
-      return {
-        ...tag,
-        commit: commit.data.commit,
-      };
-    })
+async function getTags(owner: string, repo: string) {
+  let response: GitHubGraphqlTagResponse = await graphqlWithAuth(
+    gql`
+      query GET_TAGS($owner: String!, $repo: String!) {
+        repository(owner: $owner, name: $repo) {
+          refs(
+            refPrefix: "refs/tags/"
+            first: 100
+            orderBy: { field: TAG_COMMIT_DATE, direction: DESC }
+          ) {
+            nodes {
+              name
+              target {
+                oid
+                ... on Tag {
+                  tagger {
+                    date
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    `,
+    { owner, repo }
   );
+
+  return response.repository.refs.nodes.filter((node) => {
+    return (
+      node.name.startsWith(PACKAGE_VERSION_TO_FOLLOW) ||
+      node.name.startsWith("v0.0.0-nightly-")
+    );
+  });
 }
 
 export async function getIssuesClosedByPullRequests(